Add tests to Repository class.

+++ src/ZooShop.WebSite.Home/Data/Entities/OrderEntity.cs
@@ -11,9 +11,9 @@ namespace ZooShop.Website.Home.Data.Entities
             Products = new HashSet<ProductEntity>();
         }
 
-        public long Id { get; set; }
-        public long? UserId { get; set; }
-        public byte? StatusId { get; set; }
+        public int Id { get; set; }
+        public int? UserId { get; set; }
+        public int? StatusId { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Surname { get; set; }
