+++ src/ZooShop.WebSite.Home/Data/Entities/OrderItemEntity.cs
@@ -3,9 +3,9 @@ namespace ZooShop.Website.Home.Data.Entities
 {
     public class OrderItemEntity
     {
-        public long Id { get; set; }
-        public long OrderId { get; set; }
-        public long ProductId { get; set; }
+        public int Id { get; set; }
+        public int OrderId { get; set; }
+        public int ProductId { get; set; }
         public int Quantity { get; set; }
 
         public virtual OrderEntity Order { get; set; }
