+++ src/ZooShop.WebSite.Home/Data/Entities/UserEntity.cs
@@ -14,8 +14,7 @@ namespace ZooShop.Website.Home.Data.Entities
             Products = new HashSet<ProductEntity>();
         }
 
-        public long Id { get; set; }
-        //public long CartId { get; set; }
+        public int Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Surname { get; set; }
@@ -26,7 +25,6 @@ namespace ZooShop.Website.Home.Data.Entities
 
 
         public virtual ICollection<CartItemEntity> CartItems { get; set; }
-        //public virtual Cart Cart { get; set; }
         public virtual ICollection<OrderEntity> Orders { get; set; }
         public virtual ICollection<ProductEntity> Products { get; set; }
     }
